+++ backend/cinemaBackend/routes/terms.js
@@ -1,6 +1,8 @@
 const express = require('express');
+const loginVerify = require('./verifyToken');
 const router = express.Router();
 const Term = require('../models/Term');
+const User = require('../models/User');
 
 router.get('/', async (req, res) => {
     try {
@@ -11,7 +13,7 @@ router.get('/', async (req, res) => {
     }
 });
 
-router.post('/', async (req, res) => {
+router.post('/', loginVerify, async (req, res) => {
     const term = new Term({
         date: req.body.date,
         totalPlaces: req.body.totalPlaces,
@@ -19,6 +21,10 @@ router.post('/', async (req, res) => {
     });
 
     try {
+        const user = await User.findById(req.user._id);
+        if (!user.admin) {
+            return res.status(401).send('Access only for admin');
+        }
         const savedTerm = await term.save();
         res.status(201).json(savedTerm);
     } catch(err) {
@@ -39,8 +45,12 @@ router.get('/:id', async (req, res) => {
     }
 });
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', loginVerify, async (req, res) => {
     try {
+        const user = await User.findById(req.user._id);
+        if (!user.admin) {
+            return res.status(401).send('Access only for admin');
+        }
         const updatedTerm = await Term.findByIdAndUpdate(req.params.id, req.body, {
             new: true,
             runValidators: true,
@@ -53,11 +63,32 @@ router.put('/:id', async (req, res) => {
     }
 });
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', loginVerify, async (req, res) => {
     try {
+        const user = await User.findById(req.user._id);
+        if (!user.admin) {
+            return res.status(401).send('Access only for admin');
+        }
         await Term.findByIdAndDelete(req.params.id);
         res.status(200);
         res.send();
+    } catch(err) {
+        res.status(400).json({error_message: err});
+    }
+});
+
+router.put('/:id/addreservation', loginVerify, async (req, res) => {
+    try {
+        const term = await Term.findById(req.params.id);
+        if (req.user._id != term.customer) {
+            return res.status(401).send('No access');
+        }
+        term.reservations.push({
+            customer: req.body.customer,
+            numberOfPlaces: req.body.numberOfPlaces
+        });
+        const updatedTerm = await term.save();
+        res.status(200).json(updatedTerm);
     } catch(err) {
         res.status(400).json({error_message: err});
     }
